Add files via upload
Investigating Netflix Movies and Guest Stars in The Office. The objective was to detect if duration of the movies is decreasing from the year 2011 to 2022. By analysing data we discovered that it's true.

p.s: the file with color data (color_dataset.csv) is not showing the data but it's not that important in order understand the concept of the project as data in netflix_data.csv.